requirements.txt
protobuf==4.23.0
protocol==0.1.0
requests~=2.28.2
google~=3.0.0
pynng~=0.7.2
cffi==1.15.1
APScheduler~=3.10.1
setuptools~=67.6.1
aenum~=3.1.12
pytest~=7.2.2
paramiko~=3.1.0
PyYAML~=6.0.3
requests>=2.28.0
